+++ src/main/java/org/matsim/contrib/gcs/operation/model/CarsharingOfferModel.java
@@ -9,6 +9,6 @@ import org.matsim.contrib.gcs.operation.CarsharingOperationModel;
 public interface CarsharingOfferModel extends CarsharingOperationModel {	
 	
 	public double computeRentalCost(CarsharingOffer offer, double rentalduration);
-	public ArrayList<CarsharingOffer> computeRentalOffers(double time, CarsharingDemand demand);
+	public ArrayList<CarsharingOffer> computeRentalOffers(int time, CarsharingDemand demand);
 	
 }
